Added a new constructor taking an array of floats. (i.e. a greyscale image in raw form)
Added a detectZeroCrossing method. The algorithm needs some serious work
but it does okay. (Used in LabNine)

Added threshold methods for both ints and floats.

Added a normalizeArray function that takes a float[][] and linearly
normalizes all values to the range [0,1]

Added gradientAngle, a function to get the angle of the gradient at
a specific point. This function currently takes an array of floats
instead of computing them on the fly because the assumption is that the
method will be called for many / every pixel in an image, and
re-creating the float[][] representation of the image every time would
make the whole operation O(N^4) which is just dumb. Should fix this in
the future by caching the result of getGreyscale(), etc.